+++ src/Components/Panel/index.tsx
@@ -4,7 +4,6 @@ import React, {Component} from 'react'
 // Components
 import {PanelRoot, PanelProps} from './Family/Panel'
 import {PanelHeader} from './Family/PanelHeader'
-import {PanelTitle} from './Family/PanelTitle'
 import {PanelBody} from './Family/PanelBody'
 import {PanelFooter} from './Family/PanelFooter'
 
@@ -13,7 +12,6 @@ export class Panel extends Component<PanelProps> {
 
   public static Panel = PanelRoot
   public static Header = PanelHeader
-  public static Title = PanelTitle
   public static Body = PanelBody
   public static Footer = PanelFooter
 
@@ -24,6 +22,5 @@ export class Panel extends Component<PanelProps> {
 
 export {PanelProps, PanelRef} from './Family/Panel'
 export * from './Family/PanelHeader'
-export * from './Family/PanelTitle'
 export * from './Family/PanelBody'
 export * from './Family/PanelFooter'